backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
pymongo==4.15.5
pyjwt==2.10.1
qrcode[pil]==8.2
requests==2.32.5
firebase-admin==7.1.0
bcrypt==5.0.0
python-dotenv==1.2.1
python-telegram-bot==22.5
